+++ wildshape-sheets-front/app/api/auth/[...nextauth]/route.ts
@@ -1,0 +1,5 @@
+import { authOptions } from "@/lib/auth";
+import NextAuth from "next-auth";
+
+const handler = NextAuth(authOptions);
+export { handler as GET, handler as POST };
