+++ wildshape-sheets-front/package.json
@@ -24,5 +24,8 @@
     "react-icons": "^4.10.1",
     "tailwindcss": "3.3.3",
     "typescript": "5.1.6"
+  },
+  "devDependencies": {
+    "@types/axios": "^0.14.0"
   }
 }
